[Pal/Linux-SGX] Correctly set masks in SIGSTRUCT
Previously, SIGSTRUCT.ATTRIBUTEMASK (which contains the masks for
enclave attributes such as DEBUG and for XFRM/XCR0 bits) and
SIGSTRUCT.MISCMASK (which contains the mask for Extended SSA frame
features) were incorrectly set to mirror SIGSTRUCT.ATTRIBUTES and
SIGSTRUCT.MISCSELECT.

This commit replaces this incorrect logic with hard-coded values for
SIGSTRUCT.ATTRIBUTEMASK and SIGSTRUCT.MISCMASK. In particular, all bits
in the masks are set to one (meaning "EINIT checks each attribute"). The
only exceptions currently are not-security-critical AVX and AVX512 bits
-- they are not reflected in the mask, and they are set in
SIGSTRUCT.ATTRIBUTES if the system supports them during generation of
the `.token` file.

Signed-off-by: Dmitrii Kuvaiskii <dmitrii.kuvaiskii@intel.com>